+++ sn_networking/src/cmd.rs
@@ -45,8 +45,6 @@ pub enum SwarmCmd {
         opts: DialOpts,
         sender: oneshot::Sender<Result<()>>,
     },
-    // Stop the continuous Kademlia Bootstrap process.
-    StopBootstrapping,
     // Returns all the peers from all the k-buckets from the local Routing Table.
     // This includes our PeerId as well.
     GetAllLocalPeers {
@@ -225,9 +223,6 @@ impl Debug for SwarmCmd {
             SwarmCmd::GetCloseGroupLocalPeers { key, .. } => {
                 write!(f, "SwarmCmd::GetCloseGroupLocalPeers {{ key: {:?} }}", key)
             }
-            SwarmCmd::StopBootstrapping => {
-                write!(f, "SwarmCmd::StopBootstrapping")
-            }
             SwarmCmd::GetLocalStoreCost { .. } => {
                 write!(f, "SwarmCmd::GetLocalStoreCost")
             }
@@ -350,9 +345,7 @@ impl SwarmDriver {
             } => {
                 let query_id = self.swarm.behaviour_mut().kademlia.get_record(key.clone());
 
-                if !expected_holders.is_empty() {
-                    debug!("Record {:?} with task {query_id:?} expected to be held by {expected_holders:?}", PrettyPrintRecordKey::from(&key));
-                }
+                debug!("Record {:?} with task {query_id:?} expected to be held by {expected_holders:?}", PrettyPrintRecordKey::from(&key));
 
                 if self
                     .pending_get_record
@@ -502,9 +495,6 @@ impl SwarmDriver {
                     Ok(_) => sender.send(Ok(())),
                     Err(e) => sender.send(Err(e.into())),
                 };
-            }
-            SwarmCmd::StopBootstrapping => {
-                self.bootstrap.stop_bootstrapping();
             }
             SwarmCmd::GetClosestPeersToAddressFromNetwork { key, sender } => {
                 let query_id = self